De-virtualize ArrayIdx and ArrayGet
Summary:
Modify the DictIdx and DictGet native helpers to handle mixed arrays as well, then dispatch to those helpers if we have a known TMixedArr input for ArrayIdx or ArrayGet. It's better to share these native helpers because these functions are all very hot so we can share the text itself.

arrayIdxS and arrayGetS have the same signature as dictIdxS and dictGetS, and they call them in tail-position, so the array functions should just jump into the dict functions and BOLT should arrange them right next to each other. I'll confirm that that happens with objdump.

In addition, simplify ArrayIdx and DictIdx to ArrayGet<None> and DictGetQuiet if their default argument is InitNull - by doing so, we avoid setting up two ABI registers.

Reviewed By: ricklavoie

Differential Revision: D20800310

fbshipit-source-id: 5472984935113507b2f31c75652c49faa8853ea1