Avoid thrashing s_lineCache and m_lineMap in coverage mode
Summary: Using hhprof I was able to determine that in the course of running hhvm in code coverage mode our lazy loading of the line table would cause the size of rss to balloon exponentially as the treadmill is used to free old copies of the line table. This diff introduces a per unit cache of the complete table, which is populated during code coverage mode. Additionally, the fast path cache is not updated in requests which are currently running in coverage mode.

Reviewed By: markw65

Differential Revision: D20772402

fbshipit-source-id: dc57c580b8c35784904f0095f8e9b2ad9f2bac4d