Stop generating top level code in the debugger REPL
Summary:
This diff extracts pseudomains from debugger units evaled from the REPL into top level functions. This diff also breaks the dependency the debugger had on VarEnvs, instead requiring that values be explicitly moved in and out of the evaled function as inout parameters. To properly handle exceptions the entire body of the function is surrounded in a try catch which returns exceptions to the runtime as a hidden parameter to be rethrown. Finally a sentinel object is used to mark uninit values passed into and out of the debugger function. A new bytecode was added to detect when values are unset (isset conflates unset and null values). Debugger code will look like the following:

```
function __debugger_main(inout $var1, inout $var2, ..., inout $__debugger_exn$output) {
  if ($var1 is __uinitSentinel) unset($var1);
  if ($var2 is __uinitSentinel) unset($var2);
  ...
  try {
    // user code
  } catch (Throwable $__debugger_exn$output) {
  } finally {
    if (\__systemlib\__debugger_is_uninit($var1)) $var1 = new __uninitSentinel;
    if (\__systemlib\__debugger_is_uninit($var2)) $var2 = new __uninitSentinel;
    ...
  }
}
```

While this feels somewhat hacky it allows us to avoid both a VarEnv and a pseudomain, while also not requiring a custom calling convention to pass uninit values into/out of debugger functions. The solution does require rebinding the context of the evaled function before executing it to imitate the current debugger behavior (executing a pseudomain included from within a method runs in the context of that method). We should revisit this behavior later but it isn't currently a blocker for any of the language or runtime work we currently plan to do.

Reviewed By: ricklavoie

Differential Revision: D20527312

fbshipit-source-id: ee8e4c12a4c1845b0d6cf14e60748db2c03a878b