Fix unwind.py's register restoration
Summary:
We used to rely on a two level c++ dispatch to sort out the registers
during unwinding. I fixed that a while ago in the runtime, but didn't
update unwind.py to understand the changes. The result was that gdb
simply didn't know the values of the callee saved registers when
unwinding through enterTC.

This shouldn't have mattered too much, because gcc spills all the
callee save registers into enterTC's frame (in order to use them to
hold constant values across the call into the tc, for reasons I can't
fathom). So as soon as you unwind one more level, all the registers
should be restored... except we also had a bug restoring sp. We were
restoring it to fp+16, which ends up as (char*)fp+128 since fp was
cast to a pointer to uintptr_t. This meant that the registers were
incorrectly restored in the next frame, and gdb would happily print
out incorrect values for them.

This pulls the saved registers from the stack, where they were saved
by enterTCHelper, and restores sp to the correct value.

Reviewed By: oulgen

Differential Revision: D20842216

fbshipit-source-id: b601de74f9f65c3cf568d9bb46b217f15d9dabd9