Remove legacy check for whether a type constant is abstract
Summary:
With the introduction of defaults, the syntactic check of whether a `ttc_type` is present, i.e.

```
abstract const type T = int;
                        ^^^
```

is no longer sufficient to determine if a type constant is abstract, as in the above example. I introduced a variant to properly specify the kind of the type constant. Unfortunately, there are a few places left checking the old way and are causing subtle bugs. This is one of them.

Differential Revision: D20806418

fbshipit-source-id: 4e7cfa1518b7332638fca0afe9f5e83b2c52be8f