Drop may_reenter from DecRefs
Summary:
DecRefs get moved around by delay_decrefs. may_reenter adds a killset for everything on the stack below the instruction's marker's stack offset. Since the DecRef might get moved past a store to the stack below the DecRef's stack offset, we could end up thinking that the DecRef killed the StStk. We seem to mostly get away with this because DecRefs *don't* actually re-enter (so they don't destroy what's on the stack), and because we don't run store-elim after delay_decrefs.

Fixing the stack offset for the DecRef is actually hard; setting it too deep (eg to the max of all the offsets between its original location and its new one) would mean that (if we actually re-entered) there would be garbage on the stack. And there's no simple way to know what the correct depth is (since *other* instructions might also have been moved around). Alternatively we could analyze each intervening instruction to see if it's store set interferes with the DecRef's killset, and not allowing the DecRef to move past such an instruction.

But since DecRefs aren't supposed to re-enter any more (and we have other code that makes this assumption), lets just stop marking them as re-entering.

Reviewed By: ricklavoie, oulgen

Differential Revision: D20849638

fbshipit-source-id: 4aca2ec6885b7021748e4b63ed10dd4b74d6db1e