only allow entries for open files
Summary:
I don't want caches of ASTs and TASTs and source text to grow indefinitely. Entries should be evicted.

Prior to this diff, if we got a Hover request (say) on a file that wasn't already open, then we'd add it to the list of open entries but there was never an occasion where we evicted it.

This diff instead enforces a solid rule: we only have entries for currently-open files; we create them upon DidOpen and evict them upon DidClose.

There was a bug in one of the integration tests where it was failing to send a well-formed DidOpen.

The case of CompletionResolveAtLocation is interesting. The user is in file A.php and they issue an autocomplete request. We send an autocomplete response offering the candidate "symbol Foo at file foo.php line 5 colum 3", displayed merely as "Foo" in the autocomplete dropdown. When the user hovers over Foo, then VSCode will send a CompletionResolve request to obtain docblock for that symbol Foo.

We'll need an entry for file Foo.php in order to obtain the docblock. But we're not going to add it to a permanent list of always-present entries. Instead we'll construct the entry on-the-fly, obtain the docblock, and discard it. This is fine because CompletionResolve requests happen at human timescale and there's plenty of time to fetch a file and parse it.

cc. tspence FYI since you'd been working specifically on this.

Reviewed By: arxanas

Differential Revision: D20834132

fbshipit-source-id: aebeaa7edf38d6dd7d812fcde0482632bb693774