refactor provider_utils.quarantine
Summary:
We are on our way to cleaning up quarantine!

The local memory backend no longer uses shared memory for any of its quarantineable stuff. It only uses it to store the name of the naming sqlite database, and that doesn't change and isn't subject to quarantine.

So, at last, we can factor quarantine into a sharedmem codepath and a localmem codepath.

This is nice because the nature of quarantine for the two is very different.

Note: the sharedmem push_local_stack functions never throw, and can be re-ordered.

Reviewed By: arxanas

Differential Revision: D20832355

fbshipit-source-id: b5dbb3d85a5f9d2370ddcdd4efa17f598898efd0