invalidate on File_close
Summary:
Ultimately I'm going to switch to a new theory of ctx within sIDE - where each entry's TAST is computed solely with respect to that entry on its own, and all other files as they are on disk.

But we're not there yet. We're still in a state where TASTs are sometimes computed according to all current entries of all open files. Therefore, when you close a file, it invalidates all TASTs. This is a trivial diff to reflect that fact.

Reviewed By: arxanas

Differential Revision: D20834158

fbshipit-source-id: d7603e742f3c70f596745e79802af4aa4e12b6e9