tweak signature of invalidate_tast_cache_of_entries
Summary:
My goal is for clientIdeDaemon to no longer carry around a `ctx`. Instead it will carry a list of entries for all open files.

There's one remaining place that uses `ctx`: update_naming_tables_for_changed_file. It is trivial to have it accept a list of entries instead.

Reviewed By: arxanas

Differential Revision: D20834142

fbshipit-source-id: c573fb58f408a583d7f9a105fdb22e8e261cb755