10/10 stop using server_env in clientIdeDaemon
Summary:
See diff 1/10 for an overview.

This diff completes my mission to remove server_env from clientIdeDaemon. Now, the clientIdeDaemon state only includes an explicit Naming_table.t and SearchUtils.si_env. I've done this so I can be more clear about when/why each of the things is used, and when it is updated.

Next step is to use this newfound power to implement File_close

Reviewed By: arxanas

Differential Revision: D20626726

fbshipit-source-id: 0903e78286612f36a429f1c022391bc005acfed2