don't reset ctx upon file change
Summary:
I don't understand the code that this diff deletes...

The ctx includes entries. When we receive a File_changed event, we're erasing the context that was there before.

I guess the existing "reset entries" behavior has the advantage that every file-save causes the list of entries to shrink. We don't yet handle FileClose, and so without this code the list of entries would just grow unnbounded.

And I think that reseting the context doesn't hurt too much. The TASTs stored in an entry might indeed become invalid upon file-save.

But those two things (invalidate-TAST and FileClose) should be handled properly, explicitly, not just accidentally in response to File_changed.

So I'll write this diff first, and work on adding them afterwards.

Reviewed By: arxanas

Differential Revision: D20608923

fbshipit-source-id: f60aeb4d19efe4ef9c9a8d9cc7e127efbc1c0e69