Provider_context.make_entry public
Summary:
sIDE is going to treat entries differently. It will maintain its own list of entries for each opened file, and it will cache each entry's AST and TAST.

But it won't store them in a context. It will only ever construct "singleton" contexts on the fly: when computing the TAST for file A.php, it will construct a singleton context with only an entry for A.php, and it will treat all other files as they are on disk.

This diff is a small step towards that. It allows sIDE to construct entries without them being part of a context.

Reviewed By: arxanas

Differential Revision: D20834156

fbshipit-source-id: e635b5bbfe24327f85296f1433b11382cb84285b