stop using ctx in clientIdeDaemon
Summary:
This diff is non-trivial.

Previously, ClientIdeDaemon used to carry around `ctx` as (1) a cache of ASTs and TASTs for entries, (2) a context for computing TASTs. The two roles weren't clearly separated, so sometimes we computed TASTs with respect to all entries, and sometimes we computed them only with respect to the current entry, and also we didn't get invalidation right.

This diff sets things straight. It does so by no longer carrying around `ctx`. Instead, it carries around a list open open-files with accompanying AST+TAST for role (1). And whenever it needs a context, it constructs a singleton context on the fly with just a single entry.

In this way, we're changing the behavior of the IDE. From now one there's a clear principle. Whenever you're working on a file, then this unsaved editor buffer is taken as the source of truth for this file, but disk-files are taken as the source of truth for all other files.

The actual code changes in this diff are mostly simple and mechanical. I'll draw attention to the two important places:
* Comments about invariants. This is the most important contribution.
* FileOpen/FileClose now have been changed a bit. I mostly retained existing behavior. Please don't spend too much time on them, because a future diff "only allow entries for opened files" changes their behavior significantly.

Reviewed By: arxanas

Differential Revision: D20834135

fbshipit-source-id: ac0052fc8470f514e3ea3cf6ab05aa4256d127cc