make Provider_backend.local_memory a type
Summary:
Prior to this diff, we had type
```
Provider_backend.t = | Local_memory of {...anonymous-struct...}
```
I want to write a bunch of functions which only ever operate upon the local memory backend, and I want this fact to be embodied in their type signature (rather than relying upon invariants).

This diff pulls the anonymous-struct out into a named struct, in preparation for that.

Reviewed By: arxanas

Differential Revision: D20834183

fbshipit-source-id: 572141eacc2938a44aeec1e0d929f2422cd36060