invalidate shallow decls according to entry
Summary:
There was a subtle correctness bug lurking here. I discovered it through the process of writing down decl-cache invariants and proving them correct.

There are two relevant invariants:
1. within quarantine, every shallow decl in the cache is either (1) from an entry and correct with respect to that entry context, or (2) not from an entry and is correct with respect to the file on disk.
2. outside quarantine, every shallow decl in the cache is correct with respect to the file on disk.

The first invariant is essential for our compute_tast computation to behave correctly. The second invariant isn't essential for this, but is instead the stronger "loop invariant" that we'll use to establish the first.

Imagine if we had previously been editing file B.php which was open in the editor, and had an entry, and it depended upon a shallow decl from file A.php. It will naturally read in A.php from disk and populate the shallow-decl-cache accordingly.

Now we open file A.php in the editor and make some changes, but it doesn't yet have an AST computed. We enter quarantine. At this point, the shallow decl cache would be violating the invariant!

The answer is that, upon quarantine entry, we must reparse A.php and use that newly-obtained list of symbols to know which ones to invalidate. In this way we'll satisfy the invariant 1. (we also rely upon invariant 2, to enforce invariant 1 for all files not in an entry).

And upon quarantine exit, we invalidate the same list of symbols. In this way we satisfy invariant 2.

Reviewed By: arxanas

Differential Revision: D20834128

fbshipit-source-id: ed10c25b41f64c8d35e65d71f1ea887801bd78e9