move invalidate_ctx_upon_file_change out of ClientIdeIncremental
Summary:
Trivial diff. All this does is move the function out of ClientIdeIncremental and into Provider_utils.

I rewrote the function to be a little shorter. It no longer constructs an intermediate SSet.t before invalidating. This has the effect that if a file defined a symbol twice, we used to remove it from the cache once, but now we do so twice. That has no difference.

Reviewed By: arxanas

Differential Revision: D20834195

fbshipit-source-id: 0bd554970a8c8be6d71cae8f041b459d2a0e4903