refactor type of invalidate_local_decl_cache_for_entries
Summary:
I'm working to consolidate decl-invalidation in a single place which upholds its invariants.

This diff is a trivial step along the way. The function invalidate_local_decl_cache_for_entries had a type signature which claimed that it worked for all kinds of backend, but its implementation and callers only ever applied to local-memory backend. So I'm changing its signature.

Reviewed By: arxanas

Differential Revision: D20834204

fbshipit-source-id: b072379cc3b9e70e14b8814465e34d0d56080094