invalidate some folded-decls and linearizations upon file-changed and quarantine-entry
Summary:
(This diff is non-trivial for correctness/implications standpoint, even though its code is simple).

Prior to this diff, we handled DidChange and quarantine by invalidating (1) shallow decls that were declared in the affected file, (2) folded decls that were declared in the affected file. And in DidChange we also invalidated all TASTs.

But that behavior wasn't enough to fulfill the correctness invariants. We actually need to invalidate ALL folded decls and ALL linearizations. That's what this diff does.

UPDATE:

Invalidating everything had terrible perf implications, so I'm scaling back at first, and merely extending invalidation to the linearizations. I'll need a lot more study and a future diffstack to work through the perf implications of invalidating everything.

Reviewed By: arxanas

Differential Revision: D20834133

fbshipit-source-id: 86365e3f9524e8719cb6380843cb4affc41ef070