Optimize loadzlq, foo{b,w} -> foo{b,w}m
Summary:
This sequence of opcodes from RetC of async functions was not getting fully
optimized to a simple testbim, but only to loadzlq + testbi:

  loadzlq    [%rbp + 0x18] => %130              # (16) t4:Int = LdARFlags t0:FramePtr
  andq       %131(2q), %130 => %132, %146       # (18) t6:Int = AndInt t4:Int, 2
  testq      %132, %132 => %147                 # (19) JmpNZero t6:Int -> B10

Fix this by letting Vptr from loadzlq to be used as long as we are not going to
use more than a dword.

Note that this is safe only on little endian architectures, but we are already
using the same unsafe optimization for all non-zero-extend loads.

Ideally vasm would be type safe, where dst and src widths would have to match
and explicit conversions would be necessary, but that's currently not the case
and vasm happily allows Vreg64 dst to be used as Vreg16 src.

Reviewed By: ricklavoie

Differential Revision: D20858460

fbshipit-source-id: 510e09401849a338a3f0192bafb56a0d853402e4