Back out "De-virtualize ArrayIdx and ArrayGet"
Summary:
Original commit changeset: 547298493511

This diff was meant to be a perf win, but it looks flat. The extra complexity (in particular, handling PHP arrays in ElemDict) isn't worth it for now. Perhaps as we have more array layout information (with HAM specialization), it will be, at which point I'll write a (potentially cleaner) version of it.

Reviewed By: ricklavoie

Differential Revision: D20863847

fbshipit-source-id: 4e37024b60dde790d46ce2aeacfefd3b9a7a0536