some helpers in typing_inference_env.rs
Summary:
This diff also makes structs in InferenceEnv persistent, but not InferenceEnv itself, because we estimate that whenever there is a need for bactracking, the first thing we'll do is likely modify the inference env, so we might as copy the inference env only when backtracking and not with every mutation.

For ergonomy and readability reasons, to avoid passing the arena to every single submethod that mutates, there is a pointer to the arena in every struct, similar to the rust codebase.

Reviewed By: andrewjkennedy

Differential Revision: D20834503

fbshipit-source-id: 2e1bf0a30a3ec9c98653559ff4d89ff5c294ee01