Telemetry.diff does numerical diffs
Summary:
We use Telemetry.diff to compare/merge two telemetry snapshots, so they can be more easily studied.

Here's an example diff. It uses ':prev' to indicate the previous value inline.
```
{
  "key1": 15,
  "key1:prev": 13,
  "key2": "world",
  "key2:prev": "hello"
  "key3": { },
  "key3:prev": null
}
```

I was finding the 'prev' hard to read, especially when the numbers where in the millions or billions of bytes.

This diff makes it emit the difference between numbers as a more readable "key:diff". It still uses "key:prev" for strings and other things that can't be subtracted.
```
{
  "key1": 15,
  "key1:diff": 2,
  "key2": "world",
  "key2:prev": "hello"
  "key3": { },
  "key3:prev": null
}
```

Differential Revision: D20856747

fbshipit-source-id: 36f6587a3db989b5c484493bfb96ebca618aa330