show counters for 2nd-typecheck
Summary:
Before this diff: when you do `hh --config profile_type_check_twice=true` then typing_check_service will typecheck each file twice. Also, it will gather "counters" in its telemetry - time+duration of ast-fetches, disk-cats, decl-accessors during the first typecheck.

After this diff: exactly the same, but now it also shows how those counters changed during the second typecheck.

As a reminder, we expect that the second typecheck shouldn't have to do as much work as the first, on the assumption that all decls will already be in sharedmem.

But we're still going to be spending a lot of allocations fetching the decls from sharedmem during that second typecheck. I want to get a handle on how many and how costly this remaining work is. That's what this diff provides.

Note: if you don't enable `--config profile_type_check_twice=true` then the output will be no different after this diff compared to how it was before.

Reviewed By: CatherineGasnier

Differential Revision: D20857152

fbshipit-source-id: ac0e823b63a631c75ac28f30b46aeadbfa839a1b