Hh_logger: add category list for better log filtering
Summary:
I noticed that the decl service now adds 2 lines for each decl if the min log level is set to debug. This results in lots of logs that are not interesting to me when I'm debugging something else.

Therefore, I'm adding the ability to specify a category that would act as an additional log filter.

If no category is set, which should be the majority of the cases, then only the min log level filtering applies.

I expect that we might want to add some default categories in the future via `hh.conf`.

Reviewed By: Wilfred

Differential Revision: D20858773

fbshipit-source-id: 66df1b66ba108f52dc1594b37c5ba8ce0ecbf52a