Provider_context.add_entry -> add_entry_if_missing
Summary:
This diff is non-trivial. It changes behavior.

Imagine a scenario where you have an unsaved document open and you do a hover over a symbol that's defined inside this document. ClientIdeDaemon will construct a ctx for the unsaved contents, generate its TAST, and then do ServerHover.go. But the behavior of ServerHover.go used to be to call "Provider_context.add_entry" which has the behavior that it will *replace* the entry currently in the context, with a new entry based on the disk contents.

Net effect is that if you hover over a symbol in the current file, you'll see the docblock as of what was on disk, not what was in the unsaved editor buffer.

With this diff, you now correctly see the docblock that was in the unsaved editor buffer.

I looked through the code. I didn't find a SINGLE place where the old semantics of "replace existing context entry with what I find on disk and also remove its TAST+AST caches" was the right behavior. In all cases, a better behavior is "leave the existing entry if there, and only if not should I add a new entry based on disk contents".

I've renamed the function `add_entry_if_missing` to make this clear.

Reviewed By: arxanas

Differential Revision: D20854741

fbshipit-source-id: 9055c2482a55bf4b47516fbd23a5923438406316