remove Provider_context.try_add_entry_from_disk
Summary: This diff removes try_add_entry_from_disk since it's a bit niche, used only in one place. Also I think the semantics of "do I expect to face the situation where a file on disk can't be read? how do I respond? why?" is subtle enough that it deservers to be written out in the caller.

Reviewed By: arxanas

Differential Revision: D20854718

fbshipit-source-id: 035fb8f80f62d11dd4b634f7103e176faafe4e31