Abstract frame local accesses out of iterator logic
Summary:
Convert over the iterator logic away from using raw
TypedValue* in order to abstract away the frame layout. Simply using
tv_lval as is typical causes too much register pressure on hot paths,
which leads to regressions. So, instead introduce "local_lval", which
is a tv_lval-like class pointing at a local (and only a local). It
gets away with only storing a pointer to the type, because given the
pointer to the type, we can calculate the pointer to the value. So, we
save a register, at the cost of some (hopefully cheap) arithmetic
operations. local_lval is only meant to be used for this iterator
case. Most code should just use tv_lval, which is more flexible.

Add support to ArgGroup to support tv_lval and local_lval from JITed
code.

Differential Revision: D20789869

fbshipit-source-id: a6146c39e943c49756ff73fb6ba624e06f9d2fba