Abstract away frame local accesses in interpreter builtin call logic
Summary:
The builtin call logic that the interpreter uses (in native.cpp) would
access both locals and the stack via a TypedValue*. Factor
coerceFCallArgs() into two versions, one that accesses the locals via
frame_local, and one that accesses the stack via a
TypedValue*. Likewise, have callFunc() use the appropriate access
method depending if we're in a FCallBuiltin or not.

When pushing lval arguments for the call, we need to potentially
materialize a TypedValue somewhere and pass its address (for the same
reasons as the previous CallBuiltin diff). Use space in Registers for
this.

Differential Revision: D20789137

fbshipit-source-id: 45ed6073ad1e2962d4818e39206982ef7060919a