Make memoize caches be agnostic to frame layout
Summary:
The memoize cache get/set functions receive a TypedValue* representing
the keys to perform the lookup. This works right now because both the
frame and the stack are just an array of TypedValue, so the functions
work on both. This will no longer be the case soon.  It will no longer
be possible to use the same functions to read keys from the locals
versus the stack, so split them into two variations. One that receives
the frame pointer and a start index, and one that receives a stack
pointer (similar to before). Add the necessary template abstractions
to support both cases.

For the stack case, we'll continue to just index into the given
pointer (though backwards to match the locals ordering). However, for
the frame pointer case, we'll call frame_local with the frame pointer
and index. This works right now because frame_local is cheap. If
frame_local becomes more complicated due to a more complicated layout,
this may become too expensive. In that case we can experiment with
further specializing the functions based on start offset (which would
allow us to const fold away all the offset calculations).

Adding the frame pointer and stack pointer variations caused yet a
further increase in the number of possible instantiations we need to
deal with in memo-cache.cpp. Its getting too big to deal with
manually, so use some ugly macro-ology to automate the various
instantiations. This also makes it less onerous to make future
changes. Likewise do some cleanup in irlower-intrinsics.cpp, which had
a lot of duplicated logic.

Reviewed By: mxw

Differential Revision: D20788765

fbshipit-source-id: 5224cc54df8f72a39b80cbe2b521748501839207