Last miscellaneous tv_lval changes for locals
Summary:
Last batch of miscellaneous tv_lval changes for things that could
iteract with locals. The big one is changing bytecode.cpp so that any
ipo function for a bytecode which takes a local-id receives a tv_lval
instead of a TypedValue*.

frame_local now returns a tv_lval instead of a TypedValue*. After
this, there *should* be no longer any dependency on the frame local
layout.

Differential Revision: D20803484

fbshipit-source-id: 5c14dd057a340ab141dac28ae89500b5165fc9bc