Use lvals for HHIR instructions which load/store locals
Summary:
Change over the HHIR instructions which touch locals (IE, LdLoc,
StLoc, CheckLoc, etc) to use the lval machinery instead of assuming
the frame layout. This is all straight-forward, except for LdLocAddr,
which now returns a LvalToCell instead of PtrToCell. One consequence
of that is that converting LdLocAddr to LocStkAddr is more
complicated, as they now return different types. In dce.cpp, while
doing the conversion, we need to add an additional ConvPtrToLval
instruction.

Differential Revision: D20802168

fbshipit-source-id: 1de466d87c6ad810037819e6f6184c7ec10e3c4f