Add lval support to CallBuiltin
Summary:
CallBuiltin only supports Cells and PtrToCell for its inputs. However,
it needs to support LvalToCell as well in order to support different
frame layouts.

This is somewhat straightforward, but there's one major
complication. If the builtin that we're calling takes a particular
parameter as a const Variant&, and the associated input is a
LvalToCell, we have an issue. The const Variant& should point to a
TypedValue, but there's no TypedValue to point to. In that case we
have to materialize the TypedValue (using the lval) onto the native
stack and pass a pointer to the stack location.

Add the necessary infrastructure to support this case. Since you can't
fool with the native stack directly while irlowering, instead it has
to be deferred as part of vcall/vinvoke lowering. This is where the
rest of the native stack manipulation is done.

While this is more expensive than typical parameter passing, it
shouldn't be a problem in practice. We'll only pass LvalToCell inputs
for NativeImpl, and we shouldn't have any hot of those in prod because
they'll be converted to FCallBuiltin (which uses PtrToStk and
Cells). It *will* be more problematic if we want to change the stack
layout in the future.

There was some dead code in ArgGroup, so remove that.

Reviewed By: jano

Differential Revision: D20788997

fbshipit-source-id: 6dabda648e01b2470321de99751fdeda33f48698