ArrayData::get returns TypedValue
Summary:
Modify ArrayData::get to return a TypedValue instead of tv_rval. If the value is missing, either return Uninit (if error is false) or throw (if error is true).

Rename ArrayData::rval to ArrayData::rvalVanilla, a method that can now only be called by callers that know that they have a vanilla array-like. There are only six call-sites that actually need the rval, all in the JIT. Migrate the remaining callsites to use ArrayData::get instead.

This diff required touching almost all callsites of the ArrayData getters. I found places that did an exists check prior to calling a getter, and modified them to avoid the double lookup.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D20874451

fbshipit-source-id: 5703f15664f99068b2f9cd3bf33c94ad2a067b75