avoid recursion in InstrIter
Summary:
`InstrIter` was implemented with recursive struct, if a `InstrSeq` is `Concat`, it stores `InstrInstr` of current `InstrSeq` on heap. Each call to `next`, the running cost is bound by `O(h)`, `h` is tree height. Also number of `Box::new()` eqauls to number of `InstrSeq`s in a tree.

Now it changes to using an explicit stack, it improves both `next()` cost and heap cost. Now not each call to `next` bound by `O(h)`, `InstrSeq`s in the body of an vec are just `O(1)`. Allocating heap will be at most `O(h)`, not each `InstrSeq`.

Reviewed By: losvald

Differential Revision: D20879326

fbshipit-source-id: 6a065b49f2e858515c229b0f1d4cf9026baf9362