Eliminate tv_rval::dummy
Summary:
We were using different sentinel values for ArrayData::rval and ArrayData::get. If the value was missing, ArrayData::rval returned a "null" rval that couldn't be dereferenced, while ArrayData::get returned a "dummy" rval that dereferenced to Uninit. This behavior was confusing and could lead to bugs.

Now that ArrayData::get returns a simple TypedValue, we can delete the dummy case.

Reviewed By: ricklavoie

Differential Revision: D20882840

fbshipit-source-id: bc488034ed6ad299ef5c38a021a7ef75d1ee2459