Create an apc_extend_ttl API
Summary:
One struggle we have with trying to reduce APC memory usage is that
hot keys falling out is either expensive overall due to the cache
refill cost, or risky for backends because of thundering herds as potentially
many threads all try to refill from an upstream. This API can be used to set a
low TTL and continually bump it on reads (potentially sampled). This ensures
we don't need to pay a refill cost on data that is read while allowing rarely
used data to expire out of cache.

Reviewed By: binliu19

Differential Revision: D20529392

fbshipit-source-id: b2bf0a03018eb124310c668f86af6d87a1f08c52