Apply hfsort using a call graph obtained by instrumenting optimized code via jumpstart
Summary:
This diff leverages jumpstart's mechanism for collecting profile data from
optimized code to improve the quality of the call graph used to sort functions
in the TC using hfsort.  Before this diff, such call graph was built based on
the JIT profiling code, which can be very inaccurate because function inlining
is not applied to profiling code.

The function order produced by hfsort was already being serialized along with
ProfData.  This diff moves the serialization of hfsort's output to the later
step of serializing the profile data collected for the optimized code.

The function order produced by hfsort is moved out of ProfData into a new
FuncOrder module, which also holds the functionality previously in
mcgen-order.cpp.

In order to build the call graph based on the optimized code, the following
meta-data is tracked in FuncOrder:
- for each PHP-call return address within an optimized tranlation, the
  outer-most function that such translation corresponds to;
- for each function that is optimized, the total size of the main/hot portion
  of its translations and prologues;
- a counter for each (outer-most) caller function to callee function.

In jumpstart seeder mode, the optimized prologues (OptPrologues) are
instrumented to increment the counter corresponding to the call into that
prologue.

When running retranslate-all in seeder mode, the call graph used to sort the
functions is still computed as before, and this still determines the order that
the optimized translations are emitted for profiling purposes.  Later, when
serializing the profile data collected for the optimized code, the call graph is
recomputed based on the profile data collected by instrumenting the optimized
code.  This new call graph is used to obtain a new order for the functions via a
second hfsort run.  Any function that was present in the first hfsort run and
missing in the second run is then appended to the final function order that is
serialized, in the same order produced by the first hfsort run.  This step
ensures that any function that was profiled is still recompiled in optimized
mode, even if it was inlined in all its call sites in the optimized code.  The
reason for doing so is because these functions may still be invoked by the
interpreter or live translations that will be later produced, and thus it's
important to JIT them in optimized mode.

Reviewed By: ricklavoie

Differential Revision: D20881335

fbshipit-source-id: 2f56ff3448ad2d51d502f00e86eb662e8cb7c65e