run.php: Check hhvm --version instead of --help as a sanity check
Summary:
`run.php` has a sanity-check where it runs the `$hhvm --help`
and looks for the string "Usage". This is annoying in two directions:
* it accepts nonsense like `cp`
* it adds (some) friction to HackNative's `hni` command, which
is meant to be cli-compatible with hhvm. `hni` uses `clap`, which
prints "\nUSAGE:..." for --help.

It's objectively better to use --version as a sanity check, since it
is more specific, along the lines of a User-Agent header. Just
check for a "HipHop " prefix.

Leave things alone for `hh_server` because it's --version output
is completely different, its just a hash and a date, nothing really
to filter on.

Reviewed By: alexeyt

Differential Revision: D20900948

fbshipit-source-id: 11d3cba06212426ff4f8616dc50746cb86ebd687